fixed param dim - offline RL validated

+++ maple/maple/data_management/obs_dict_replay_buffer.py
@@ -72,7 +72,7 @@ class ObsDictRelabelingBuffer(ReplayBuffer):
         self.ob_spaces = self.env.observation_space.spaces
         for key in self.ob_keys_to_save + internal_keys:
             assert key in self.ob_spaces, "Key not found in the observation space: %s" % key
-            type = np.float64
+            type = np.float32
             if key.startswith("image"):
                 type = np.uint8
             self._obs[key] = np.zeros((max_size, self.ob_spaces[key].low.size), dtype=type)
@@ -282,7 +282,7 @@ def postprocess_obs_dict(obs_dict):
 
 def normalize_image(image):
     assert image.dtype == np.uint8
-    return np.float64(image) / 255.0
+    return np.float32(image) / 255.0
 
 
 def unnormalize_image(image):
